requirements.txt
webdriver_manager==3.8.3
selenium==4.4.0
vtt-to-srt3==0.1.9.1